* Updated README.md code example with error control * ResultError has been moved out to it's own file * Updated docstring documentation
