+++ README.md
@@ -2,9 +2,9 @@
 
 [![Build Status](https://travis-ci.org/QISKit/qiskit-sdk-py.svg?branch=master)](https://travis-ci.org/QISKit/qiskit-sdk-py)
 
-The Quantum Information Software Kit (**QISKit** for short) is a software development kit (SDK) for working with [OpenQASM](https://github.com/QISKit/qiskit-openqasm) and the [IBM Q experience (QX)](https://quantumexperience.ng.bluemix.net/). 
+The Quantum Information Software Kit (**QISKit** for short) is a software development kit (SDK) for working with [OpenQASM](https://github.com/QISKit/qiskit-openqasm) and the [IBM Q experience (QX)](https://quantumexperience.ng.bluemix.net/).
 
-Use **QISKit** to create quantum computing programs, compile them, and execute them on one of several backends (online Real quantum processors, online simulators, and local simulators). For the online backends, QISKit uses our [python API client](https://github.com/QISKit/qiskit-api-py) to connect to the IBM Q experience. 
+Use **QISKit** to create quantum computing programs, compile them, and execute them on one of several backends (online Real quantum processors, online simulators, and local simulators). For the online backends, QISKit uses our [python API client](https://github.com/QISKit/qiskit-api-py) to connect to the IBM Q experience.
 
 **We use GitHub issues for tracking requests and bugs. Please see the** [IBM Q experience community](https://quantumexperience.ng.bluemix.net/qx/community) **for questions and discussion.** **If you'd like to contribute to QISKit, please take a look at our** [contribution guidelines](CONTRIBUTING.rst).
 
@@ -33,7 +33,7 @@ For those more familiar with python, the fastest way to install QISKit is by usi
     pip install qiskit
 ```
 
-### Source Installation 
+### Source Installation
 
 An alternative method is to clone the QISKit SDK repository onto your local machine, and change into the cloned directory:
 
@@ -41,7 +41,7 @@ An alternative method is to clone the QISKit SDK repository onto your local mach
 
 Select the "Clone or download" button at the top of this webpage (or from the URL shown in the git clone command), unzip the file if needed, and change into **qiskit-sdk-py folder** in a terminal window.
 
-#### Git download 
+#### Git download
 
 Or, if you have Git installed, run the following commands:
 
@@ -68,34 +68,40 @@ from qiskit import QuantumProgram
 # Creating Programs create your first QuantumProgram object instance.
 Q_program = QuantumProgram()
 
-# Creating Registers create your first Quantum Register called "qr" with 2 qubits
-qr = Q_program.create_quantum_register("qr", 2)
-# create your first Classical Register called "cr" with 2 bits
-cr = Q_program.create_classical_register("cr", 2)
-# Creating Circuits create your first Quantum Circuit called "qc" involving your Quantum Register "qr" 
-# and your Classical Register "cr"
-qc = Q_program.create_circuit("superposition", [qr], [cr])
+try:
+  # Creating Registers create your first Quantum Register called "qr" with 2 qubits
+  qr = Q_program.create_quantum_register("qr", 2)
+  # create your first Classical Register called "cr" with 2 bits
+  cr = Q_program.create_classical_register("cr", 2)
+  # Creating Circuits create your first Quantum Circuit called "qc" involving your Quantum Register "qr"
+  # and your Classical Register "cr"
+  qc = Q_program.create_circuit("superposition", [qr], [cr])
 
-# add the H gate in the Qubit 0, we put this Qubit in superposition
-qc.h(qr[0])
+  # add the H gate in the Qubit 0, we put this Qubit in superposition
+  qc.h(qr[0])
 
-# add measure to see the state
-qc.measure(qr, cr)
+  # add measure to see the state
+  qc.measure(qr, cr)
 
-# Compiled  and execute in the local_qasm_simulator
+  # Compiled  and execute in the local_qasm_simulator
 
-result = Q_program.execute(["superposition"], backend='local_qasm_simulator', shots=1024)
+  result = Q_program.execute(["superposition"], backend='local_qasm_simulator', shots=1024)
 
-# Show the results
-print(result)
-print(result.get_data("superposition"))
+  # Show the results
+  print(result)
+  print(result.get_data("superposition"))
+
+except QISKitError as ex:
+  print('There was an error in the circuit!. Error = {}'.format(ex))
+except RegisterSizeError as ex:
+  print('Error in the number of registers!. Error = {}'.format(ex))
 ```
 
 In this case, the output will be (approximately due to random fluctuations):
 
 ```
 COMPLETED
-{'00': 509, '11': 515} 
+{'00': 509, '11': 515}
 ```
 
 You can also use QISKit to execute your code on a [real Quantum Chip](https://github.com/QISKit/ibmqx-backend-information).
