+++ qiskit/__init__.py
@@ -1,3 +1,4 @@
+from IBMQuantumExperience import RegisterSizeError
 from ._qiskiterror import QISKitError
 from ._classicalregister import ClassicalRegister
 from ._quantumregister import QuantumRegister
@@ -10,7 +11,7 @@ import qiskit.extensions.standard
 from ._jobprocessor import JobProcessor
 from ._quantumjob import QuantumJob
 from ._quantumprogram import QuantumProgram
-from ._result import (Result, ResultError)
-from IBMQuantumExperience import RegisterSizeError
+from ._result import Result
+
 
 __version__ = '0.4.0'
