+++ qiskit/_result.py
@@ -1,6 +1,7 @@
 import copy
 
 from qiskit._qiskiterror import QISKitError
+from qiskit import RegisterSizeError
 
 class Result(object):
     """ Result Class.
@@ -140,10 +141,16 @@ class Result(object):
 
         Returns:
             A dictionary of data for the different backends.
-            If there's an error, the function will throw.
+
+        Raises:
+            If there's an error the function will throw a QISKitError or a
+            RegisterSizeError.
         """
         if self._is_error():
-            raise self.__result['result'] # Rethrow exception!
+            exception = self.__result['result']
+            if isinstance(exception, RegisterSizeError):
+                raise exception # Re-throw RegisterSizeError
+            raise QISKitError(str(exception))
 
         try:
             qobj = self.__qobj
@@ -201,22 +208,4 @@ class Result(object):
             if key in observable:
                 temp += counts[key] * observable[key] / tot
         return temp
-
-
-class ResultError(QISKitError):
-    """Exceptions raised due to errors in result output.
-
-    It may be better for the QISKit API to raise this exception.
-
-    Args:
-        error (dict): This is the error record as it comes back from
-            the API. The format is like::
-
-                error = {'status': 403,
-                         'message': 'Your credits are not enough.',
-                         'code': 'MAX_CREDITS_EXCEEDED'}
-    """
-    def __init__(self, error):
-        self.status = error['status']
-        self.message = error['message']
-        self.code = error['code']
+
